Delete Confirmation for Appointments complete

+++ src/main/java/cordinus/cordinus_global/controller/ApptController.java
@@ -232,24 +232,28 @@ public class ApptController {
                                 ResultSet rs = ps.executeQuery();
 
                                 if(rs.next()){
-
-                                        int x = rs.getInt(1);
-                                        String y = rs.getString(5);
+                                        //int x = rs.getInt(1);
+                                        //String y = rs.getString(5);
                                         Alert alert = new Alert(Alert.AlertType.WARNING);
                                         alert.setTitle("Delete Warning");
                                         alert.setHeaderText("Deleting Appointment");
-                                        alert.setContentText("Are you sure you wish to delete this appointment?\nAPPOINTMENT_TYPE : " + y + "\nAPPOINTMENT_ID# : " + x);
+                                        alert.setContentText("Are you sure you wish to delete this appointment?\nAPPOINTMENT_TYPE : " +
+                                                AppointmentTable.getSelectionModel().getSelectedItem().getType() + "\nAPPOINTMENT_ID# : " +
+                                                AppointmentTable.getSelectionModel().getSelectedItem().getAppointment_ID());
                                         Optional<ButtonType> result = alert.showAndWait();
                                         if(result.isPresent()  && result.get() != ButtonType.CANCEL){
-                                                AppointmentsQuery.delete(AppointmentTable.getSelectionModel().getSelectedItem().getAppointment_ID());
-                                        }else{
-                                                Alert delete_alert = new Alert(Alert.AlertType.CONFIRMATION);
-                                                delete_alert.setTitle("Delete Confirmation");
-                                                delete_alert.setHeaderText("Appointment Deleted");
-                                                delete_alert.setContentText("APPOINTMENT_TYPE : "
-                                                        + y + "APPOINTMENT_ID# : " + x +" has been canceled.");
-                                        }
+                                                if(result.isPresent()  && result.get() ==ButtonType.OK){
+                                                        Alert delete_alert = new Alert(Alert.AlertType.CONFIRMATION);
+                                                        delete_alert.setTitle("Delete Confirmation");
+                                                        delete_alert.setHeaderText("Appointment Deleted");
+                                                        delete_alert.setContentText("\nAPPOINTMENT_TYPE : "
+                                                                + AppointmentTable.getSelectionModel().getSelectedItem().getType() +
+                                                                "\nAPPOINTMENT_ID# : " + AppointmentTable.getSelectionModel().getSelectedItem().getAppointment_ID() +" has been deleted.");
+                                                        Optional<ButtonType> delete_result = delete_alert.showAndWait();
+                                                        AppointmentsQuery.delete(AppointmentTable.getSelectionModel().getSelectedItem().getAppointment_ID());
+                                                }
 
+                                        }
                                 }
 
                         }catch(NullPointerException e){
